Add Board.hash method and fix contract tests
- Implement hash() method in Board interface
- Update test contracts and mock implementations accordingly
- Import test contracts from saumarai package

+++ src/main/java/com/joansala/engine/doe/DOENode.java
@@ -143,7 +143,7 @@ public class DOENode implements Comparable<DOENode>, Serializable {
     /**
      * Number of expanded childs.
      */
-    public double getCount() {
+    public long getCount() {
         return count;
     }
